wip: merge put files into one file
